服务端口自动生成 Consul宕机恢复侦测 查询服务信息使用缓冲

+++ shell/StartConsul.sh
@@ -2,6 +2,9 @@
 CURRDIR=`pwd`
 cd ../../../../..
 export GOPATH=`pwd`
-go install -v github.com/hashicorp/consul
+if [ ! -f "${GOPATH}/bin/consul" ]; then
+	go install -v github.com/hashicorp/consul
+fi
+
 cd ${GOPATH}/bin
 ./consul agent -server -bind=127.0.0.1 -bootstrap-expect=1  -client=127.0.0.1 -ui -data-dir=.
